requirements.txt
langchain==0.0.354
langchain-community==0.0.8
langchain-core==0.1.5
langsmith==0.0.77
Pillow==10.2.0
streamlit==1.29.0
openai==0.28.0
toml==0.10.2